factor out the part of types that doesn't use pydantic
