+++ python/cog/types_v1.py
@@ -10,14 +10,7 @@ from typing import Any, Dict, Iterator, List, Optional, TypeVar, Union
 
 import requests
 from pydantic import Field, SecretStr
-
-FILENAME_ILLEGAL_CHARS = set("\u0000/")
-
-# Linux allows files up to 255 bytes long. We enforce a slightly shorter
-# filename so that there's room for prefixes added by
-# tempfile.NamedTemporaryFile, etc.
-FILENAME_MAX_LENGTH = 200
-
+from .types_shared import URLFile, URLPath, get_filename
 
 def Input(
     default: Any = ...,
@@ -112,140 +105,6 @@ class Path(pathlib.PosixPath):
         field_schema.update(type="string", format="uri")
 
 
-class URLPath(pathlib.PosixPath):
-    """
-    URLPath is a nasty hack to ensure that we can defer the downloading of a
-    URL passed as a path until later in prediction dispatch.
-
-    It subclasses pathlib.PosixPath only so that it can pass isinstance(_,
-    pathlib.Path) checks.
-    """
-
-    _path: Optional[Path]
-
-    def __init__(self, *, source: str, filename: str, fileobj: io.IOBase) -> None:
-        self.source = source
-        self.filename = filename
-        self.fileobj = fileobj
-
-        self._path = None
-
-    def convert(self) -> Path:
-        if self._path is None:
-            dest = tempfile.NamedTemporaryFile(suffix=self.filename, delete=False)
-            shutil.copyfileobj(self.fileobj, dest)
-            self._path = Path(dest.name)
-        return self._path
-
-    def unlink(self, missing_ok: bool = False) -> None:
-        if self._path:
-            # TODO: use unlink(missing_ok=...) when we drop Python 3.7 support.
-            try:
-                self._path.unlink()
-            except FileNotFoundError:
-                if not missing_ok:
-                    raise
-
-    def __str__(self) -> str:
-        # FastAPI's jsonable_encoder will encode subclasses of pathlib.Path by
-        # calling str() on them
-        return self.source
-
-
-class URLFile(io.IOBase):
-    """
-    URLFile is a proxy object for a :class:`urllib3.response.HTTPResponse`
-    object that is created lazily. It's a file-like object constructed from a
-    URL that can survive pickling/unpickling.
-    """
-
-    __slots__ = ("__target__", "__url__")
-
-    def __init__(self, url: str) -> None:
-        object.__setattr__(self, "__url__", url)
-
-    # We provide __getstate__ and __setstate__ explicitly to ensure that the
-    # object is always picklable.
-    def __getstate__(self) -> Dict[str, Any]:
-        return {"url": object.__getattribute__(self, "__url__")}
-
-    def __setstate__(self, state: Dict[str, Any]) -> None:
-        object.__setattr__(self, "__url__", state["url"])
-
-    # Proxy getattr/setattr/delattr through to the response object.
-    def __setattr__(self, name: str, value: Any) -> None:
-        if hasattr(type(self), name):
-            object.__setattr__(self, name, value)
-        else:
-            setattr(self.__wrapped__, name, value)
-
-    def __getattr__(self, name: str) -> Any:
-        if name in ("__target__", "__wrapped__", "__url__"):
-            raise AttributeError(name)
-        else:
-            return getattr(self.__wrapped__, name)
-
-    def __delattr__(self, name: str) -> None:
-        if hasattr(type(self), name):
-            object.__delattr__(self, name)
-        else:
-            delattr(self.__wrapped__, name)
-
-    # Luckily the only dunder method on HTTPResponse is __iter__
-    def __iter__(self) -> Iterator[bytes]:
-        return iter(self.__wrapped__)
-
-    @property
-    def __wrapped__(self) -> Any:
-        try:
-            return object.__getattribute__(self, "__target__")
-        except AttributeError:
-            url = object.__getattribute__(self, "__url__")
-            resp = requests.get(url, stream=True)
-            resp.raise_for_status()
-            resp.raw.decode_content = True
-            object.__setattr__(self, "__target__", resp.raw)
-            return resp.raw
-
-    def __repr__(self) -> str:
-        try:
-            target = object.__getattribute__(self, "__target__")
-        except AttributeError:
-            return "<{} at 0x{:x} for {!r}>".format(
-                type(self).__name__, id(self), object.__getattribute__(self, "__url__")
-            )
-        else:
-            return f"<{type(self).__name__} at 0x{id(self):x} wrapping {target!r}>"
-
-
-def get_filename(url: str) -> str:
-    parsed_url = urllib.parse.urlparse(url)
-
-    if parsed_url.scheme == "data":
-        resp = urllib.request.urlopen(url)  # noqa: S310
-        mime_type = resp.headers.get_content_type()
-        extension = mimetypes.guess_extension(mime_type)
-        if extension is None:
-            return "file"
-        return "file" + extension
-
-    basename = os.path.basename(parsed_url.path)
-    basename = urllib.parse.unquote_plus(basename)
-
-    # If the filename is too long, we truncate it (appending '~' to denote the
-    # truncation) while preserving the file extension.
-    # - truncate it
-    # - append a tilde
-    # - preserve the file extension
-    if _len_bytes(basename) > FILENAME_MAX_LENGTH:
-        basename = _truncate_filename_bytes(basename, length=FILENAME_MAX_LENGTH)
-
-    for c in FILENAME_ILLEGAL_CHARS:
-        basename = basename.replace(c, "_")
-
-    return basename
-
-
 Item = TypeVar("Item")
 
 
@@ -270,17 +129,3 @@ class ConcatenateIterator(Iterator[Item]):
     @classmethod
     def validate(cls, value: Iterator[Any]) -> Iterator[Any]:
         return value
-
-
-def _len_bytes(s: str, encoding: str = "utf-8") -> int:
-    return len(s.encode(encoding))
-
-
-def _truncate_filename_bytes(s: str, length: int, encoding: str = "utf-8") -> str:
-    """
-    Truncate a filename to at most `length` bytes, preserving file extension
-    and avoiding text encoding corruption from truncation.
-    """
-    root, ext = os.path.splitext(s.encode(encoding))
-    root = root[: length - len(ext) - 1]
-    return root.decode(encoding, "ignore") + "~" + ext.decode(encoding)
